Separate out instrumentation for netty 4.0 from 4.1
