+++ dd-java-agent/instrumentation/netty-4.0/src/main/java/datadog/trace/instrumentation/netty40/client/HttpClientTracingHandler.java
@@ -1,0 +1,17 @@
+package datadog.trace.instrumentation.netty40.client;
+
+import io.netty.channel.CombinedChannelDuplexHandler;
+import io.netty.util.AttributeKey;
+import io.opentracing.Span;
+
+public class HttpClientTracingHandler
+    extends CombinedChannelDuplexHandler<
+        HttpClientResponseTracingHandler, HttpClientRequestTracingHandler> {
+
+  static final AttributeKey<Span> attributeKey =
+      new AttributeKey<>(HttpClientTracingHandler.class.getName());
+
+  public HttpClientTracingHandler() {
+    super(new HttpClientResponseTracingHandler(), new HttpClientRequestTracingHandler());
+  }
+}
