+++ dd-java-agent/instrumentation/netty-4.1/src/main/java/datadog/trace/instrumentation/netty41/NettyChannelPipelineInstrumentation.java
@@ -1,0 +1,128 @@
+package datadog.trace.instrumentation.netty41;
+
+import static net.bytebuddy.matcher.ElementMatchers.hasSuperType;
+import static net.bytebuddy.matcher.ElementMatchers.isInterface;
+import static net.bytebuddy.matcher.ElementMatchers.isMethod;
+import static net.bytebuddy.matcher.ElementMatchers.nameStartsWith;
+import static net.bytebuddy.matcher.ElementMatchers.named;
+import static net.bytebuddy.matcher.ElementMatchers.not;
+import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
+
+import com.google.auto.service.AutoService;
+import datadog.trace.agent.tooling.DDAdvice;
+import datadog.trace.agent.tooling.DDTransformers;
+import datadog.trace.agent.tooling.HelperInjector;
+import datadog.trace.agent.tooling.Instrumenter;
+import datadog.trace.bootstrap.CallDepthThreadLocalMap;
+import datadog.trace.instrumentation.netty41.client.HttpClientRequestTracingHandler;
+import datadog.trace.instrumentation.netty41.client.HttpClientResponseTracingHandler;
+import datadog.trace.instrumentation.netty41.client.HttpClientTracingHandler;
+import datadog.trace.instrumentation.netty41.server.HttpServerRequestTracingHandler;
+import datadog.trace.instrumentation.netty41.server.HttpServerResponseTracingHandler;
+import datadog.trace.instrumentation.netty41.server.HttpServerTracingHandler;
+import io.netty.channel.ChannelHandler;
+import io.netty.channel.ChannelPipeline;
+import io.netty.handler.codec.http.HttpClientCodec;
+import io.netty.handler.codec.http.HttpRequestDecoder;
+import io.netty.handler.codec.http.HttpRequestEncoder;
+import io.netty.handler.codec.http.HttpResponseDecoder;
+import io.netty.handler.codec.http.HttpResponseEncoder;
+import io.netty.handler.codec.http.HttpServerCodec;
+import net.bytebuddy.agent.builder.AgentBuilder;
+import net.bytebuddy.asm.Advice;
+
+@AutoService(Instrumenter.class)
+public class NettyChannelPipelineInstrumentation extends Instrumenter.Configurable {
+
+  private static final String PACKAGE =
+      NettyChannelPipelineInstrumentation.class.getPackage().getName();
+
+  public NettyChannelPipelineInstrumentation() {
+    super("netty", "netty-4.1");
+  }
+
+  @Override
+  protected boolean defaultEnabled() {
+    return false;
+  }
+
+  @Override
+  public AgentBuilder apply(final AgentBuilder agentBuilder) {
+    return agentBuilder
+        .type(not(isInterface()).and(hasSuperType(named("io.netty.channel.ChannelPipeline"))))
+        .transform(
+            new HelperInjector(
+                // client helpers
+                PACKAGE + ".client.NettyResponseInjectAdapter",
+                PACKAGE + ".client.HttpClientRequestTracingHandler",
+                PACKAGE + ".client.HttpClientResponseTracingHandler",
+                PACKAGE + ".client.HttpClientTracingHandler",
+                // server helpers
+                PACKAGE + ".server.NettyRequestExtractAdapter",
+                PACKAGE + ".server.HttpServerRequestTracingHandler",
+                PACKAGE + ".server.HttpServerResponseTracingHandler",
+                PACKAGE + ".server.HttpServerTracingHandler"))
+        .transform(DDTransformers.defaultTransformers())
+        .transform(
+            DDAdvice.create()
+                .advice(
+                    isMethod()
+                        .and(nameStartsWith("add"))
+                        .and(takesArgument(2, named("io.netty.channel.ChannelHandler"))),
+                    ChannelPipelineAddAdvice.class.getName()))
+        .asDecorator();
+  }
+
+  /**
+   * When certain handlers are added to the pipeline, we want to add our corresponding tracing
+   * handlers. If those handlers are later removed, we may want to remove our handlers. That is not
+   * currently implemented.
+   */
+  public static class ChannelPipelineAddAdvice {
+    @Advice.OnMethodEnter
+    public static int checkDepth() {
+      return CallDepthThreadLocalMap.incrementCallDepth(ChannelPipeline.class);
+    }
+
+    @Advice.OnMethodExit(suppress = Throwable.class)
+    public static void addHandler(
+        @Advice.Enter final int depth,
+        @Advice.This final ChannelPipeline pipeline,
+        @Advice.Argument(2) final ChannelHandler handler) {
+      if (depth > 0) return;
+
+      try {
+        // Server pipeline handlers
+        if (handler instanceof HttpServerCodec) {
+          pipeline.addLast(
+              HttpServerTracingHandler.class.getName(), new HttpServerTracingHandler());
+        } else if (handler instanceof HttpRequestDecoder) {
+          pipeline.addLast(
+              HttpServerRequestTracingHandler.class.getName(),
+              new HttpServerRequestTracingHandler());
+        } else if (handler instanceof HttpResponseEncoder) {
+          pipeline.addLast(
+              HttpServerResponseTracingHandler.class.getName(),
+              new HttpServerResponseTracingHandler());
+        } else
+        // Client pipeline handlers
+        if (handler instanceof HttpClientCodec) {
+          pipeline.addLast(
+              HttpClientTracingHandler.class.getName(), new HttpClientTracingHandler());
+        } else if (handler instanceof HttpRequestEncoder) {
+          pipeline.addLast(
+              HttpClientRequestTracingHandler.class.getName(),
+              new HttpClientRequestTracingHandler());
+        } else if (handler instanceof HttpResponseDecoder) {
+          pipeline.addLast(
+              HttpClientResponseTracingHandler.class.getName(),
+              new HttpClientResponseTracingHandler());
+        }
+      } catch (final IllegalArgumentException e) {
+        // Prevented adding duplicate handlers.
+      } finally {
+        CallDepthThreadLocalMap.reset(ChannelPipeline.class);
+      }
+    }
+  }
+}
