+++ dd-java-agent/instrumentation/netty-4.1/src/main/java/datadog/trace/instrumentation/netty41/server/HttpServerTracingHandler.java
@@ -1,0 +1,17 @@
+package datadog.trace.instrumentation.netty41.server;
+
+import io.netty.channel.CombinedChannelDuplexHandler;
+import io.netty.util.AttributeKey;
+import io.opentracing.Span;
+
+public class HttpServerTracingHandler
+    extends CombinedChannelDuplexHandler<
+        HttpServerRequestTracingHandler, HttpServerResponseTracingHandler> {
+
+  static final AttributeKey<Span> attributeKey =
+      AttributeKey.valueOf(HttpServerTracingHandler.class.getName());
+
+  public HttpServerTracingHandler() {
+    super(new HttpServerRequestTracingHandler(), new HttpServerResponseTracingHandler());
+  }
+}
