--- dd-java-agent/instrumentation/netty/netty.gradle
@@ -1,19 +1,0 @@
-apply from: "${rootDir}/gradle/java.gradle"
-
-
-dependencies {
-  compileOnly group: 'io.netty', name: 'netty-all', version: '4.1.25.Final'
-
-  compile project(':dd-java-agent:agent-tooling')
-
-  compile deps.bytebuddy
-  compile deps.opentracing
-  annotationProcessor deps.autoservice
-  implementation deps.autoservice
-
-  testCompile project(':dd-java-agent:testing')
-
-  testCompile group: 'org.eclipse.jetty', name: 'jetty-server', version: '9.4.11.v20180605'
-  testCompile group: 'com.squareup.okhttp3', name: 'okhttp', version: '3.6.0'
-  testCompile group: 'org.asynchttpclient', name: 'async-http-client', version: '2.4.9'
-}
