--- dd-java-agent/instrumentation/netty/src/main/java/datadog/trace/instrumentation/netty/NettyChannelPipelineInstrumentation.java
@@ -1,125 +1,0 @@
-package datadog.trace.instrumentation.netty;
-
-import static net.bytebuddy.matcher.ElementMatchers.hasSuperType;
-import static net.bytebuddy.matcher.ElementMatchers.isInterface;
-import static net.bytebuddy.matcher.ElementMatchers.isMethod;
-import static net.bytebuddy.matcher.ElementMatchers.nameStartsWith;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.not;
-import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
-
-import com.google.auto.service.AutoService;
-import datadog.trace.agent.tooling.DDAdvice;
-import datadog.trace.agent.tooling.DDTransformers;
-import datadog.trace.agent.tooling.HelperInjector;
-import datadog.trace.agent.tooling.Instrumenter;
-import datadog.trace.bootstrap.CallDepthThreadLocalMap;
-import datadog.trace.instrumentation.netty.client.HttpClientRequestTracingHandler;
-import datadog.trace.instrumentation.netty.client.HttpClientResponseTracingHandler;
-import datadog.trace.instrumentation.netty.client.HttpClientTracingHandler;
-import datadog.trace.instrumentation.netty.server.HttpServerRequestTracingHandler;
-import datadog.trace.instrumentation.netty.server.HttpServerResponseTracingHandler;
-import datadog.trace.instrumentation.netty.server.HttpServerTracingHandler;
-import io.netty.channel.ChannelHandler;
-import io.netty.channel.ChannelPipeline;
-import io.netty.handler.codec.http.HttpClientCodec;
-import io.netty.handler.codec.http.HttpRequestDecoder;
-import io.netty.handler.codec.http.HttpRequestEncoder;
-import io.netty.handler.codec.http.HttpResponseDecoder;
-import io.netty.handler.codec.http.HttpResponseEncoder;
-import io.netty.handler.codec.http.HttpServerCodec;
-import net.bytebuddy.agent.builder.AgentBuilder;
-import net.bytebuddy.asm.Advice;
-
-@AutoService(Instrumenter.class)
-public class NettyChannelPipelineInstrumentation extends Instrumenter.Configurable {
-
-  public NettyChannelPipelineInstrumentation() {
-    super("netty", "netty-http-server");
-  }
-
-  @Override
-  protected boolean defaultEnabled() {
-    return false;
-  }
-
-  @Override
-  public AgentBuilder apply(final AgentBuilder agentBuilder) {
-    return agentBuilder
-        .type(not(isInterface()).and(hasSuperType(named("io.netty.channel.ChannelPipeline"))))
-        .transform(
-            new HelperInjector(
-                // client helpers
-                "datadog.trace.instrumentation.netty.client.NettyResponseInjectAdapter",
-                "datadog.trace.instrumentation.netty.client.HttpClientRequestTracingHandler",
-                "datadog.trace.instrumentation.netty.client.HttpClientResponseTracingHandler",
-                "datadog.trace.instrumentation.netty.client.HttpClientTracingHandler",
-                // server helpers
-                "datadog.trace.instrumentation.netty.server.NettyRequestExtractAdapter",
-                "datadog.trace.instrumentation.netty.server.HttpServerRequestTracingHandler",
-                "datadog.trace.instrumentation.netty.server.HttpServerResponseTracingHandler",
-                "datadog.trace.instrumentation.netty.server.HttpServerTracingHandler"))
-        .transform(DDTransformers.defaultTransformers())
-        .transform(
-            DDAdvice.create()
-                .advice(
-                    isMethod()
-                        .and(nameStartsWith("add"))
-                        .and(takesArgument(2, named("io.netty.channel.ChannelHandler"))),
-                    ChannelPipelineAddAdvice.class.getName()))
-        .asDecorator();
-  }
-
-  /**
-   * When certain handlers are added to the pipeline, we want to add our corresponding tracing
-   * handlers. If those handlers are later removed, we may want to remove our handlers. That is not
-   * currently implemented.
-   */
-  public static class ChannelPipelineAddAdvice {
-    @Advice.OnMethodEnter
-    public static int checkDepth() {
-      return CallDepthThreadLocalMap.incrementCallDepth(ChannelPipeline.class);
-    }
-
-    @Advice.OnMethodExit(suppress = Throwable.class)
-    public static void addHandler(
-        @Advice.Enter final int depth,
-        @Advice.This final ChannelPipeline pipeline,
-        @Advice.Argument(2) final ChannelHandler handler) {
-      if (depth > 0) return;
-
-      try {
-        // Server pipeline handlers
-        if (handler instanceof HttpServerCodec) {
-          pipeline.addLast(
-              HttpServerTracingHandler.class.getName(), new HttpServerTracingHandler());
-        } else if (handler instanceof HttpRequestDecoder) {
-          pipeline.addLast(
-              HttpServerRequestTracingHandler.class.getName(),
-              new HttpServerRequestTracingHandler());
-        } else if (handler instanceof HttpResponseEncoder) {
-          pipeline.addLast(
-              HttpServerResponseTracingHandler.class.getName(),
-              new HttpServerResponseTracingHandler());
-        } else
-        // Client pipeline handlers
-        if (handler instanceof HttpClientCodec) {
-          pipeline.addLast(
-              HttpClientTracingHandler.class.getName(), new HttpClientTracingHandler());
-        } else if (handler instanceof HttpRequestEncoder) {
-          pipeline.addLast(
-              HttpClientRequestTracingHandler.class.getName(),
-              new HttpClientRequestTracingHandler());
-        } else if (handler instanceof HttpResponseDecoder) {
-          pipeline.addLast(
-              HttpClientResponseTracingHandler.class.getName(),
-              new HttpClientResponseTracingHandler());
-        }
-      } catch (final IllegalArgumentException e) {
-        // Prevented adding duplicate handlers.
-      } finally {
-        CallDepthThreadLocalMap.reset(ChannelPipeline.class);
-      }
-    }
-  }
-}
