--- dd-java-agent/instrumentation/netty/src/main/java/datadog/trace/instrumentation/netty/client/HttpClientTracingHandler.java
@@ -1,12 +1,0 @@
-package datadog.trace.instrumentation.netty.client;
-
-import io.netty.channel.CombinedChannelDuplexHandler;
-
-public class HttpClientTracingHandler
-    extends CombinedChannelDuplexHandler<
-        HttpClientResponseTracingHandler, HttpClientRequestTracingHandler> {
-
-  public HttpClientTracingHandler() {
-    super(new HttpClientResponseTracingHandler(), new HttpClientRequestTracingHandler());
-  }
-}
