--- dd-java-agent/instrumentation/netty/src/main/java/datadog/trace/instrumentation/netty/server/HttpServerTracingHandler.java
@@ -1,12 +1,0 @@
-package datadog.trace.instrumentation.netty.server;
-
-import io.netty.channel.CombinedChannelDuplexHandler;
-
-public class HttpServerTracingHandler
-    extends CombinedChannelDuplexHandler<
-        HttpServerRequestTracingHandler, HttpServerResponseTracingHandler> {
-
-  public HttpServerTracingHandler() {
-    super(new HttpServerRequestTracingHandler(), new HttpServerResponseTracingHandler());
-  }
-}
