+++ gradle/java.gradle
@@ -50,8 +50,10 @@ dependencies {
   testCompile deps.groovy
   testCompile deps.testLogging
   testCompile 'info.solidsoft.spock:spock-global-unroll:0.5.1'
-  testCompile group: 'io.ratpack', name: 'ratpack-groovy-test', version: '1.4.6'
   testCompile group: 'com.github.stefanbirkner', name: 'system-rules', version: '1.17.1'
+  if (!project.name.contains("netty")) { // avoid screwing up the classpath...
+    testCompile group: 'io.ratpack', name: 'ratpack-groovy-test', version: '1.4.6'
+  }
 }
 
 tasks.withType(Javadoc) {
