+++ settings.gradle
@@ -36,7 +36,8 @@ include ':dd-java-agent:instrumentation:kafka-streams-0.11'
 include ':dd-java-agent:instrumentation:lettuce-5'
 include ':dd-java-agent:instrumentation:mongo-3.1'
 include ':dd-java-agent:instrumentation:mongo-async-3.3'
-include ':dd-java-agent:instrumentation:netty'
+include ':dd-java-agent:instrumentation:netty-4.0'
+include ':dd-java-agent:instrumentation:netty-4.1'
 include ':dd-java-agent:instrumentation:okhttp-3'
 include ':dd-java-agent:instrumentation:osgi-classloading'
 include ':dd-java-agent:instrumentation:play-2.4'
